backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
celery[redis]==5.4.0
redis==5.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==43.0.1
python-multipart==0.0.12
openai==1.51.0
dnspython==2.7.0
openpyxl==3.1.5
aiofiles==24.1.0
websockets==13.1
httpx==0.27.2
